requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
exceptiongroup==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
mysql-connector-python==9.1.0
mysqlclient==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
